create a derived class EditEmployeePanel
